AC: remove anchor TS dependency from SDK
AC: fix unused, missing imports

AC: remove unused anchor imports

AC: remove test utils
