+++ account-compression/Anchor.toml
@@ -16,4 +16,4 @@ cluster = "localnet"
 wallet = "~/.config/solana/id.json"
 
 [scripts]
-test = "yarn run ts-mocha -t 1000000 tests/**/*-test.ts"
+test = "yarn run ts-mocha -t 1000000 tests/**/*.test.ts"
