+++ account-compression/programs/spl-compression/src/state/mod.rs
@@ -2,5 +2,5 @@
 mod concurrent_merkle_tree_header;
 mod path_node;
 
-pub use concurrent_merkle_tree_header::{ConcurrentMerkleTreeHeader, CompressionAccountType};
+pub use concurrent_merkle_tree_header::ConcurrentMerkleTreeHeader;
 pub use path_node::PathNode;
