+++ account-compression/sdk/package.json
@@ -21,8 +21,8 @@
     "build": "rm -rf dist/ && tsc && tsc -p tsconfig.cjs.json"
   },
   "dependencies": {
-    "@project-serum/anchor": "0.24.2",
     "@solana/web3.js": "^1.50.1",
+    "bn.js": "^5.2.1",
     "borsh": "^0.7.0",
     "chai": "^4.3.4"
   },
