+++ account-compression/tests/merkleTree.ts
@@ -1,10 +1,7 @@
-import { BN } from "@project-serum/anchor";
 import { keccak_256 } from "js-sha3";
 import * as Collections from 'typescript-collections';
 import { PublicKey } from "@solana/web3.js";
-import { fetch } from 'cross-fetch';
 
-const MAX_DEPTH = 20;
 let CACHE_EMPTY_NODE = new Map<number, Buffer>();
 
 export type Tree = {
@@ -29,6 +26,9 @@ export function getRoot(tree: Tree): TreeNode {
     return node;
 }
 
+/*
+ Breadth-first iteration over a merkle tree
+*/
 export function bfs<T>(tree: Tree, iterFunc: (node: TreeNode, nodeIdx: number) => T): T[] {
     let toExplore = [getRoot(tree)];
     const results = []
@@ -50,15 +50,6 @@ export function bfs<T>(tree: Tree, iterFunc: (node: TreeNode, nodeIdx: number) =
     }
     return results;
 }
-
-const generateLeafNode = (seeds) => {
-    let leaf = Buffer.alloc(32);
-    for (const seed of seeds) {
-        leaf = Buffer.from(keccak_256.digest([...leaf, ...seed]));
-    }
-    return leaf;
-};
-
 
 export function emptyNode(level: number): Buffer {
     if (CACHE_EMPTY_NODE.has(level)) {
@@ -208,82 +199,6 @@ export function updateTree(tree: Tree, newNode: Buffer, index: number, verbose =
  */
 export function hash(left: Buffer, right: Buffer): Buffer {
     return Buffer.from(keccak_256.digest(Buffer.concat([left, right])));
-}
-
-/**
- *  Does not build tree, just returns root of tree from leaves
- */
-export function hashLeaves(leaves: Buffer[]): Buffer {
-    let nodes = leaves;
-    let level = 0;
-    while (level < MAX_DEPTH) {
-        let next_nodes = [];
-
-        if (nodes.length == 0) {
-            nodes = [emptyNode(level), emptyNode(level)];
-        }
-
-        while (nodes.length > 0) {
-            let left = nodes.pop();
-            let right: Buffer;
-
-            if (nodes.length > 0) {
-                right = nodes.pop();
-            } else {
-                right = emptyNode(level);
-            }
-            next_nodes.push(hash(left, right));
-        }
-
-        level++;
-        nodes = next_nodes
-    }
-    return nodes[0];
-}
-
-
-export type AssetProof = {
-    root: string,
-    hash: string,
-    proof: string[],
-}
-
-export type NodeView = {
-    hash: string,
-    level: number,
-    index: number
-}
-
-// Very annoying, since last element in proof is actually root
-export async function getProofOfLeafFromServer(endpoint: string, treeId: PublicKey, i: number): Promise<NodeView[]> {
-    const url = `${endpoint}/proof/${treeId.toString()}/${i}`;
-    let object = await fetch(
-        url,
-        {
-            method: "GET",
-        }
-    ).then(resp => resp.json());
-
-    return object.data as NodeView[];
-}
-
-export async function getProofOfAssetFromServer(endpoint: string, treeId: PublicKey, i: number): Promise<AssetProof> {
-    const url = `${endpoint}/assets/${treeId.toString()}/${i}/proof`;
-    let object = await fetch(
-        url,
-        {
-            method: "GET",
-        }
-    ).then(resp => resp.json())
-        .catch((e) => console.log(e));
-
-    return object.data as AssetProof;
-}
-
-export async function getRootFromServer(endpoint: string, treeId: PublicKey): Promise<string> {
-    const url = `${endpoint}/root/${treeId.toString()}`;
-    let object = await fetch(url, { method: "GET", }).then(resp => resp.json());
-    return object.data as string;
 }
 
 export function checkProof(index: number, root: string, leaf: string, proof: string[], verbose = false) {
